tutorials/pubsub-cloudrun-pipeline/load-records-to-pubsub/requirements.txt
Flask==2.0.2
gunicorn==20.1.0
cloudevents
google-cloud-storage
google-cloud-pubsub
